uint for shard and nonce

+++ src/Mx.NET.SDK/Domain/Data/Blocks/Block.cs
@@ -50,7 +50,7 @@ namespace Mx.NET.SDK.Domain.Data.Blocks
         /// <summary>
         /// Block shard
         /// </summary>
-        public long Shard { get; private set; }
+        public uint Shard { get; private set; }
 
         /// <summary>
         /// Block size
